requirements.txt
django
gunicorn
django-heroku
django-environ
djangorestframework